feat: use ktor networking library for fetching config SDKCF-1276
- configure project to support multiplatform
- move fetcher into common source
