+++ build.gradle
@@ -5,6 +5,9 @@ buildscript {
   CONFIG.versions.kotlin = '1.3.21'
   CONFIG.versions.android.plugin = '3.3.1'
   CONFIG.versions.okhttp = '3.12.3'
+  CONFIG.versions.ktor_version = '1.1.2'
+  CONFIG.kotlinx_coroutines_version = '1.1.1'
+  CONFIG.serialization_version = '0.10.0'
 
   repositories {
     google()
@@ -31,6 +34,7 @@ allprojects { proj ->
     google()
     jcenter()
     maven { url "https://kotlin.bintray.com/kotlinx" }
+    maven { url 'https://dl.bintray.com/kotlin/ktor' }
   }
 
   apply plugin: 'pl.allegro.tech.build.axion-release'
