+++ remote-config/build.gradle
@@ -1,8 +1,8 @@
 apply from: '../config/android/library.gradle'
-apply plugin: "kotlin-android"
 apply plugin: 'kotlin-android-extensions'
-apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlinx-serialization'
+apply plugin: 'org.jetbrains.kotlin.multiplatform'
+apply plugin: 'kotlin-kapt'
 
 android {
   sourceSets.each {
@@ -16,6 +16,31 @@ android {
   }
 }
 
+kotlin {
+    targets {
+        fromPreset(presets.android, 'androidLib')
+    }
+
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation 'org.jetbrains.kotlin:kotlin-stdlib-common'
+                implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core-common:$CONFIG.kotlinx_coroutines_version"
+                implementation "io.ktor:ktor-client-core:$CONFIG.versions.ktor_version"
+                implementation "io.ktor:ktor-client-json:$CONFIG.versions.ktor_version"
+            }
+        }
+
+        androidLibMain {
+            dependencies {
+                implementation "com.squareup.okhttp3:okhttp:$CONFIG.versions.okhttp"
+                implementation "io.ktor:ktor-client-android:$CONFIG.versions.ktor_version"
+                implementation "io.ktor:ktor-client-json-jvm:$CONFIG.versions.ktor_version"
+            }
+        }
+    }
+}
+
 dependencies {
   implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$CONFIG.versions.kotlin"
   implementation "com.squareup.okhttp3:okhttp:$CONFIG.versions.okhttp"
@@ -23,15 +48,14 @@ dependencies {
   implementation "com.rakuten.tech.mobile:manifest-config-annotations:0.1.0"
   kapt "com.rakuten.tech.mobile:manifest-config-processor:0.1.0"
   implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime:0.11.0"
-  implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.1.1"
 
-  testImplementation 'junit:junit:4.12'
-  testImplementation 'org.robolectric:robolectric:3.8'
-  testImplementation "org.amshove.kluent:kluent-android:1.48"
-  testImplementation "com.nhaarman.mockitokotlin2:mockito-kotlin:2.1.0"
-  testImplementation "org.mockito:mockito-inline:2.28.2"
-  testImplementation "com.squareup.okhttp3:mockwebserver:$CONFIG.versions.okhttp"
-  testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.0-M1'
+  androidTestImplementation 'junit:junit:4.12'
+  androidTestImplementation 'org.robolectric:robolectric:3.8'
+  androidTestImplementation "org.amshove.kluent:kluent-android:1.48"
+  androidTestImplementation "com.nhaarman.mockitokotlin2:mockito-kotlin:2.1.0"
+  androidTestImplementation "org.mockito:mockito-inline:2.28.2"
+  androidTestImplementation "com.squareup.okhttp3:mockwebserver:$CONFIG.versions.okhttp"
+  //androidTestImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.0-M1'
 }
 
 apply from: '../gradle/documentation.gradle'
