+++ settings.gradle
@@ -1,1 +1,3 @@
+enableFeaturePreview('GRADLE_METADATA')
+
 include ':remote-config', ':sample'
